[SDR-357] refactor: Feed, ReviewDetail 중복 코드 제거
- FeedCard 컴포넌트 생성 및 사용
